linprocfs: Fix i386 time type
Cast the time type to (long). This is correct on all architectures. On
i386, this promotes the int time_t to a long (which is also 32-bit). On
64-bit architectures, this promotes the 64-bit signed time_t to a 64-bit
signed int type.

Sponsored by:		Netflix

+++ sys/compat/linprocfs/linprocfs.c
@@ -2145,9 +2145,9 @@ again:
 			    msqids[id].u.msg_perm.gid,
 			    msqids[id].u.msg_perm.cuid,
 			    msqids[id].u.msg_perm.cgid,
-			    msqids[id].u.msg_stime,
-			    msqids[id].u.msg_rtime,
-			    msqids[id].u.msg_ctime);
+			    (long)msqids[id].u.msg_stime,
+			    (long)msqids[id].u.msg_rtime,
+			    (long)msqids[id].u.msg_ctime);
 
 	free(msqids, M_TEMP);
 	return (0);
@@ -2199,8 +2199,8 @@ again:
 			    semids[id].u.sem_perm.gid,
 			    semids[id].u.sem_perm.cuid,
 			    semids[id].u.sem_perm.cgid,
-			    semids[id].u.sem_otime,
-			    semids[id].u.sem_ctime);
+			    (long)semids[id].u.sem_otime,
+			    (long)semids[id].u.sem_ctime);
 
 	free(semids, M_TEMP);
 	return (0);
@@ -2256,9 +2256,9 @@ again:
 			    shmids[id].u.shm_perm.gid,
 			    shmids[id].u.shm_perm.cuid,
 			    shmids[id].u.shm_perm.cgid,
-			    shmids[id].u.shm_atime,
-			    shmids[id].u.shm_dtime,
-			    shmids[id].u.shm_ctime,
+			    (long)shmids[id].u.shm_atime,
+			    (long)shmids[id].u.shm_dtime,
+			    (long)shmids[id].u.shm_ctime,
 			    0, 0);	/* XXX rss & swp are not supported */
 
 	free(shmids, M_TEMP);
